tests/wpt/web-platform-tests/2dcontext/text-styles/2d.text.font.parse.invalid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.text.font.parse.invalid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>2d.text.font.parse.invalid</h1>
<p class="desc"></p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

ctx.font = '20px serif';
_assertSame(ctx.font, '20px serif', "ctx.font", "'20px serif'");

ctx.font = '20px serif';
ctx.font = 'bogus';
_assertSame(ctx.font, '20px serif', "ctx.font", "'20px serif'");

ctx.font = '20px serif';
ctx.font = 'inherit';
_assertSame(ctx.font, '20px serif', "ctx.font", "'20px serif'");

ctx.font = '20px serif';
ctx.font = '10px {bogus}';
_assertSame(ctx.font, '20px serif', "ctx.font", "'20px serif'");

ctx.font = '20px serif';
ctx.font = '10px initial';
_assertSame(ctx.font, '20px serif', "ctx.font", "'20px serif'");

ctx.font = '20px serif';
ctx.font = '10px default';
_assertSame(ctx.font, '20px serif', "ctx.font", "'20px serif'");

ctx.font = '20px serif';
ctx.font = '10px inherit';
_assertSame(ctx.font, '20px serif', "ctx.font", "'20px serif'");

ctx.font = '20px serif';
ctx.font = '1em serif; background: green; margin: 10px';
_assertSame(ctx.font, '20px serif', "ctx.font", "'20px serif'");


});
</script>
